requirements.txt
pinecone
groq
langchain
langchain-pinecone
langchain-huggingface
langchain-community
pypdf
python-dotenv
sentence-transformers